+++ server/routes/roundRouter.js
@@ -1,15 +1,11 @@
 const express = require('express')
 const {
-    getLeagueGradeSeason,
     ensureAuthenticated,
     ensureLeagueAdmin,
 } = require('./utils')
 const roundController = require('../controllers/roundController.js')
 
 const roundRouter = express.Router()
-
-// GET
-roundRouter.get('/:roundId', getLeagueGradeSeason, roundController.getRound)
 
 // POST
 roundRouter.post(
